Added the -H option to run jobs on bluefire.
git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@5880 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: cb8c33a3e8e80932c51e9776fdd497bf57b3b359

+++ postprocessing/output_scripts/voca/exec_voca_output_creater.bash
@@ -6,10 +6,13 @@ myhelp()
 { 
 echo "------------------------------------------------------------------------------------"
 echo "This bash script was created to run the matlab script voca_output_creater.m."
-echo "The Matlab script has to be located in the same directory as this script. If"
-echo "you have large input files the -b option might be very usefull. The matlab"
-echo "script will be running in the background then and you don't have to stay"
-echo "logged in. For more Information have a look at ticket 47 of the WRF trac."
+echo "The Matlab script has to be located in the same directory as this script. To"
+echo "run jobs on HD1 use the -H option. If you have large input files the -b "
+echo "option might be very usefull. The matlab script will be running in the "
+echo "background then and you don't have to stay logged in. For more Information"
+echo "have a look at ticket 47 of the WRF trac."
+echo ""
+echo "Dependencies: matlab, [atd, bsub]"
 echo "" 
 echo "Usage: ./exec_voca_output_creater.bash [-bm] -a <action> -f <infile> | -d <indir>"
 echo ""
@@ -29,8 +32,9 @@ indir=''
 action=''
 background=false
 merge=false
+hd1=false
 
-while getopts hbma:f:d: opt
+while getopts hHbma:f:d: opt
 do
   case "$opt" in
     f) infile="$OPTARG";;
@@ -38,8 +42,9 @@ do
     d) indir="$OPTARG";;
     b) background=true;;
     m) merge=true;;
+    H) hd1=true;;
     h) myhelp; exit;;
-    \?) echo "Error: Unknown option $opt."; myhelp;;
+    \?) echo "Error: Unknown option."; myhelp;;
   esac
 done
 
@@ -63,18 +68,45 @@ elif [ "$infile" = "" ] && [ "$indir" = "" ]; then
 	exit
 fi
 
+if [ $background = true ] && [ $hd1 = true ]; then
+	echo "Error: You cannot use the -b and -H option at the same time!"
+	exit
+fi
+
 # if an existing input file was specified
 if [ "$infile" != "" ] && [ -f $infile ]; then
 
+	#run in background
 	if [ $background = true ]; then
 		echo "matlab -r voca_output_creater\(\'$infile\'\,\'$action\'\)" >tmp
+
+		# merge eps to pdf
 		if [ $merge = true ]; then	
 			echo "./mergeEPS.bash -f ${infile%/*}/ ${infile%/*}/plots.pdf">>tmp
 		fi 
+
 		at now -f tmp
 		rm tmp
+
+	#run in foreground
 	else
-		matlab -r voca_output_creater\(\'$infile\'\,\'$action\'\)
+		# hd1 run
+		if [ $hd1 = true ]; then
+			if [ -f run_matlab.job ]; then
+				rm run_matlab.job
+			fi
+			touch run_matlab.job
+			echo '#BSUB -J MatlabJob'>>run_matlab.job
+			echo '#BSUB -o MatlabJob_output'>>run_matlab.job
+			echo '/sharedapps/LS/matlab/bin/matlab -nodisplay -nosplash -nojvm -r '"voca_output_creater\(\'$infile\'\,\'$action\'\)">>run_matlab.job
+			bsub < run_matlab.job
+
+		# usual run		
+		else
+			matlab -r voca_output_creater\(\'$infile\'\,\'$action\'\)
+		fi		
+			
+		# merge eps to pdf
 		if [ $merge = true ]; then	
 			./mergeEPS.bash -f ${infile%/*}/ ${infile%/*}/plots.pdf
 		fi 
@@ -83,7 +115,7 @@ if [ "$infile" != "" ] && [ -f $infile ]; then
 # if an existing input directory was specified
 elif [ "$indir" != "" ] && [ -d $indir ]; then
 	k=0	
-	#get files from indir; exclude all png and eps
+	# get files from indir; exclude all png and eps
 	for curFile in $indir*; do
 		if [ "${curFile/*./}" != "png" ] && [ "${curFile/*./}" != "eps" ]; then
 			files[$k]=$curFile;
@@ -99,16 +131,38 @@ elif [ "$indir" != "" ] && [ -d $indir ]; then
 	done
 	fileList=$fileList'\}'
 
+	# run in background
 	if [ $background = true ]; then
 		echo "matlab -r voca_output_creater\($fileList\,\'$action\'\)" >tmp
+		
+		# merge eps to pdf
 		if [ $merge = true ]; then	
 			echo "./mergeEPS.bash -f ${infile%/*}/ ${infile%/*}/plots.pdf">>tmp
-		fi 
+		fi
+ 
 		at now -f tmp
 		rm tmp
+
+	# run in foreground
 	else		
-		echo "matlab -r voca_output_creater\($fileList\,\'$action\'\)"
-		matlab -r voca_output_creater\($fileList\,\'$action\'\)
+		# hd1 run
+		if [ $hd1 = true ]; then
+			if [ -f run_matlab.job ]; then
+				rm run_matlab.job
+			fi
+			touch run_matlab.job
+			echo '#BSUB -J MatlabJob'>>run_matlab.job
+			echo '#BSUB -o MatlabJob_output'>>run_matlab.job
+			echo '/sharedapps/LS/matlab/bin/matlab -nodisplay -nosplash -nojvm -r '"voca_output_creater\($fileList\,\'$action\'\)">>run_matlab.job
+			bsub < run_matlab.job
+
+		# usual run
+		else 
+			echo "matlab -r voca_output_creater\($fileList\,\'$action\'\)"
+			matlab -r voca_output_creater\($fileList\,\'$action\'\)
+		fi		
+
+		# merge eps to pdf
 		if [ $merge = true ]; then	
 			./mergeEPS.bash -f ${infile%/*}/ ${infile%/*}/plots.pdf
 		fi
